Fixing a small re-entracy bug in the fetch API which lead to a crash when deleting cache items.

+++ Classes/ISCacheItem.h
@@ -57,7 +57,7 @@ static const int ISCacheItemTotalBytesUnknown = -1;
 @property (strong) NSDate *modified;
 @property (strong) NSError *lastError;
 
-@property (readonly) CGFloat progress;
+@property (readonly) float progress;
 @property (readonly) NSTimeInterval timeRemainingEstimate;
 
 + (id)itemWithIdentifier:(NSString *)identifier
